last modified

+++ src/users/users.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Body, Injectable, NotFoundException, Req, UnauthorizedException } from '@nestjs/common';
+import { BadRequestException, Body, ForbiddenException, Injectable, NotFoundException, Req, UnauthorizedException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Users } from './user.entity';
 import { Repository } from 'typeorm';
@@ -21,15 +21,25 @@ export class UsersService {
         private readonly departmentRepository: Repository<Department>
     ) { }
 
-    async findOneByEmail(email: string): Promise<Users | null> {
-        return this.usersRepository
-            .createQueryBuilder('user')
-            .leftJoinAndSelect('user.company', 'company')
-            .where('user.email = :email', { email })
-            .getOne();
+    async createUser(dto: CreateUserDto, adminId: number): Promise<Users | null> {
+        const admin = await this.usersRepository.findOne({
+            where: { id: adminId },
+            relations: ['company']
+        });
+
+        if (!admin || !admin.company) {
+            throw new UnauthorizedException('Admin has no company assigned')
+        }
+
+        const newUser = this.usersRepository.create({
+            ...dto,
+            company: admin.company,
+        })
+
+        return this.usersRepository.save(newUser);
     }
 
-    async findAll(adminId: number) {
+    async findAll(adminId: number): Promise<Users[]> {
         const admin = await this.usersRepository.findOne({
             where: { id: adminId },
             relations: ['company']
@@ -43,26 +53,80 @@ export class UsersService {
         })
 
         return users;
-
     }
 
-    async findById(id: number): Promise<Users | null> {
-        return this.usersRepository.findOne({
-            where: { id },
-            relations: ['company', 'department']
+    async findOne(adminId: number, id: number): Promise<Users | null> {
+        const admin = await this.usersRepository.findOne({
+            where: { id: adminId },
+            relations: ['company']
         });
+
+        if (!admin || !admin.company) throw new UnauthorizedException('Admin has no company assigned');
+
+        const user = await this.usersRepository.findOne({
+            where: { id },
+            relations: ['company']
+        })
+
+        if (user?.company?.id !== admin.company.id) {
+            throw new ForbiddenException("Something went wrong")
+        }
+
+        return user
     }
 
-    async update(id: number, updateUserDto: UpdateUserDto): Promise<Users | null> {
+    async findOneByEmail(email: string): Promise<Users | null> {
+        return this.usersRepository
+            .createQueryBuilder('user')
+            .leftJoinAndSelect('user.company', 'company')
+            .where('user.email = :email', { email })
+            .getOne();
+    }
+
+    async update(adminId: number, id: number, updateUserDto: UpdateUserDto): Promise<Users | null> {
+        const admin = await this.usersRepository.findOne({
+            where: { id: adminId },
+            relations: ['company']
+        })
+
+        if (!admin || !admin.company) throw new UnauthorizedException('Admin has no company assigned')
+
+        const user = await this.usersRepository.findOne({
+            where: { id },
+            relations: ['company']
+        })
+
+        if (user?.company?.id !== admin.company.id) {
+            throw new ForbiddenException("Something went wrong")
+        }
+
         await this.usersRepository.update(id, updateUserDto);
-        return this.findOne(id);
-    }
-
-    findOne(id: number): Promise<Users | null> {
         return this.usersRepository.findOne({ where: { id } });
     }
 
-    async remove(id: number): Promise<void> {
+    async delete(adminId: number, id: number): Promise<void> {
+        const admin = await this.usersRepository.findOne({
+            where: { id: adminId },
+            relations: ['company']
+        });
+
+        if (!admin || !admin.company) {
+            throw new UnauthorizedException('Admin has no company assigned');
+        }
+
+        const user = await this.usersRepository.findOne({
+            where: { id },
+            relations: ['company']
+        })
+
+        if (!user) {
+            throw new NotFoundException('User not found')
+        }
+
+        if (user?.company?.id !== admin.company.id) {
+            throw new ForbiddenException("Something went wrong")
+        }
+
         await this.usersRepository.delete(id);
     }
 
@@ -74,35 +138,7 @@ export class UsersService {
         return this.usersRepository.save(user);
     }
 
-    async createFromAdmin(dto: CreateUserDto, adminId: number) {
-        const admin = await this.usersRepository.findOne({
-            where: { id: adminId },
-            relations: ['company']
-        });
 
-        if (!admin || !admin.company) throw new UnauthorizedException('Admin has no company assigned')
-
-        const department = await this.departmentRepository.findOne({
-            where: {
-                id: dto.department_id,
-                company: { id: admin.company.id },
-            },
-            relations: ['company'],
-        });
-
-        if (!department) {
-            throw new BadRequestException('Invalid department selected');
-        }
-
-        const newUser = this.usersRepository.create({
-            ...dto,
-            company: admin.company,
-            department: department,
-        })
-
-        return this.usersRepository.save(newUser);
-
-    }
 
 
 }
